MU/requirements.txt
torch
transformers
datasets
tokenizers
numpy
scipy
matplotlib
seaborn
pandas
scikit-learn
nltk
jupyter
pytest
pyyaml
tqdm
